Final Project Day 8 (7/12)
The following changes were made:

1. New types of rectangles were added

a.  A trampoline-like rectangle which lets the circle bounce twice as high (a separate method was created for this)

b. A stationary rectangle which kills the circle upon landing

c. A moving rectangle which kills the circle falls right through it

e. A rectangle which breaks/disappears when the circle bounces on it
This rectangle is still buggy and while a new method was created for it, it's not fully coded yet.