Final Project (8/12)
The following changes were made:

1. New types of rectangle were added:

a. A trampoline-like rectangle (green) which lets the circle bounce twice as high (a separate method was created for this and the process is much smoother now)

b. A stationary rectangle (black) which kills the circle upon landing

c. A moving rectangle (dark red) which kills the circle upon landing

d. A trap rectangle (white) which doesn't let the circle bounce

e. A rectangle which breaks/disappears when the circle bounces on it (brown)
This rectangle is still buggy, so I've commented it all out for now

2. The circle now doesn't leave the screen while jumping upwards (it stops  after reaching a certain height)

3. I tried displaying and changing the score based on the number of times the ball bounces but this is still buggy so, I've commented it all out

4. The rectangles now respawn only in the upper half of the screen once they go underneath the screen and disappear